% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIUQ-package.R
\docType{package}
\name{AIUQ-package}
\alias{AIUQ}
\alias{AIUQ-package}
\title{AIUQ: Ab Initio Uncertainty Quantification}
\description{
Uncertainty quantification and inverse estimation by probabilistic generative models from the beginning of the data analysis. An example is a Fourier basis method for inverse estimation in scattering analysis of microscopy videos. It does not require specifying a certain range of Fourier bases and it substantially reduces computational cost via the generalized Schur algorithm. See the reference: Mengyang Gu, Yue He, Xubo Liu and Yimin Luo (2023), \doi{10.48550/arXiv.2309.02468}.
}
\author{
\strong{Maintainer}: Mengyang Gu \email{mengyang@pstat.ucsb.edu}

Authors:
\itemize{
  \item Yue He
  \item Xubo Liu
}

}

\keyword{internal}

