% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corr_fBM}
\alias{corr_fBM}
\title{Construct correlation matrix for FBM}
\usage{
corr_fBM(len_t, H)
}
\arguments{
\item{len_t}{number of time steps}

\item{H}{Hurst parameter, value between 0 and 1}
}
\value{
Correlation matrix with dimension \code{len_t-1} by \code{len_t-1}.
}
\description{
Construct correlation matrix for fractional Brownian motion.
}
\examples{
library(AIUQ)
len_t = 50
H = 0.3
m = corr_fBM(len_t=len_t,H=H)
}
\author{
\packageAuthor{AIUQ}
}
\keyword{internal}
