% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_grad_trans}
\alias{get_grad_trans}
\title{Construct parameter transformation for optimization using exact gradient}
\usage{
get_grad_trans(theta, d_input, model_name)
}
\arguments{
\item{theta}{parameters to be optimized over}

\item{d_input}{sequence of lag times}

\item{model_name}{model name for the fitted model, options from ('BM','OU',
'FBM',OU+FBM','user_defined')}
}
\value{
A vector of transformed parameters to be optimized over in \code{AIUQ}
method of \code{SAM} class.
}
\description{
Construct parameter transformation for parameters to be optimized over in \code{AIUQ}
method of \code{SAM} class. See 'References'.
}
\references{
Gu, M., He, Y., Liu, X., & Luo, Y. (2023). Ab initio uncertainty
quantification in scattering analysis of microscopy.
arXiv preprint arXiv:2309.02468.

Gu, M., Luo, Y., He, Y., Helgeson, M. E., & Valentine, M. T. (2021).
Uncertainty quantification and estimation in differential dynamic microscopy.
Physical Review E, 104(3), 034610.
}
\author{
\packageAuthor{AIUQ}
}
\keyword{internal}
