% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_traj}
\alias{plot_traj}
\title{Plot 2D particle trajectory}
\usage{
plot_traj(object, title = NA)
}
\arguments{
\item{object}{an S4 object of class \code{simulation}}

\item{title}{main title of the plot. If \code{NA}, title is "model_name with
M particles" with \code{model_name} and \code{M} being field in \code{simulation}
class.}
}
\value{
2D plot of particle trajectory for a given simulation from \code{simulation}
class.
}
\description{
Function to plot the particle trajectory after the \code{simulation} class
has been constructed.
}
\examples{
library(AIUQ)
sim_bm = simulation(sz=100,len_t=100,sigma_bm=0.5)
show(sim_bm)
plot_traj(sim_bm)
}
\author{
\packageAuthor{AIUQ}
}
